requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
